+++ locale/opsiPackageBuilder_de.ts
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE TS><TS version="2.0" language="de_DE" sourcelanguage="en">
+<!DOCTYPE TS>
+<TS version="2.1" language="de_DE" sourcelanguage="en">
 <context>
     <name>BaseController</name>
     <message>
@@ -782,12 +783,12 @@ Folgende Ausgabe(n) wurde(n) zurückgegeben:</translation>
         <translation>Das Paket wurde bereits erstellt und wird nicht überschrieben!</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="826"/>
+        <location filename="../oPB/core/processing.py" line="868"/>
         <source>Command not found. See Log for details.</source>
         <translation>Befehl nicht gefunden. Siehe Logbuch für Details.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="842"/>
+        <location filename="../oPB/core/processing.py" line="884"/>
         <source>Error establishing SSH connection. See Log for details.</source>
         <translation>Fehler beim Verbindungsaufbau. Siehe Logbuch für Details.</translation>
     </message>
@@ -797,27 +798,27 @@ Folgende Ausgabe(n) wurde(n) zurückgegeben:</translation>
         <translation>Die Paketdatei konnte nicht gefunden werden!</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="884"/>
+        <location filename="../oPB/core/processing.py" line="926"/>
         <source>There are umlauts in some fields which can&apos;t be processed: build failed.</source>
         <translation>Nicht verarbeitbare Umlaute gefunden, Paketbau abgebrochen.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="888"/>
+        <location filename="../oPB/core/processing.py" line="930"/>
         <source>Backend error: installation failed.</source>
         <translation>Fehler im Backend: Installation abgebrochen.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="892"/>
+        <location filename="../oPB/core/processing.py" line="934"/>
         <source>Could not process &apos;install&apos; command.</source>
         <translation>Konnte das Kommando &apos;install&apos; nicht verarbeiten.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="896"/>
+        <location filename="../oPB/core/processing.py" line="938"/>
         <source>Package file error. Check log.</source>
         <translation>Paketdatei fehlerhaft. Bitte das Logbuch prüfen.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="908"/>
+        <location filename="../oPB/core/processing.py" line="962"/>
         <source>Undefined error occurred. Check log.</source>
         <translation>Unspezifizierter Fehler aufgetreten. Bitte das Logbuch prüfen.</translation>
     </message>
@@ -827,7 +828,7 @@ Folgende Ausgabe(n) wurde(n) zurückgegeben:</translation>
         <translation>Fehler während der Befehlsausführung. Für Details Logbuch prüfen.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="713"/>
+        <location filename="../oPB/core/processing.py" line="731"/>
         <source>opsi-product-updater is already running.</source>
         <translation>Der opsi-product-updater läuft bereits.</translation>
     </message>
@@ -842,14 +843,29 @@ Folgende Ausgabe(n) wurde(n) zurückgegeben:</translation>
         <translation>Extrahieren...</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="900"/>
+        <location filename="../oPB/core/processing.py" line="942"/>
         <source>Error during package extraction. Check log.</source>
         <translation>Fehler beim Entpacken des Pakets. Für Details Logbuch prüfen.</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="904"/>
+        <location filename="../oPB/core/processing.py" line="946"/>
         <source>Permission denied. Check log.</source>
         <translation>Erlaubnis verweigert. Für Details Logbuch prüfen.</translation>
+    </message>
+    <message>
+        <location filename="../oPB/core/processing.py" line="950"/>
+        <source>Network logon failed. Check log.</source>
+        <translation>Netzwerk Logon nicht möglich. Für Details Logbuch prüfen.</translation>
+    </message>
+    <message>
+        <location filename="../oPB/core/processing.py" line="954"/>
+        <source>Possible error: command or file not found. Check log.</source>
+        <translation>Möglicher Fehler: Befehl oder Datei nicht gefunden. Für Details Logbuch prüfen.</translation>
+    </message>
+    <message>
+        <location filename="../oPB/core/processing.py" line="958"/>
+        <source>Failed to get ip address for host. Check log.</source>
+        <translation>Konnte IP Adresse des Hosts nicht ermitteln. Für Details Logbuch prüfen.</translation>
     </message>
 </context>
 <context>
@@ -896,7 +912,7 @@ Folgende Ausgabe(n) wurde(n) zurückgegeben:</translation>
         <translation type="obsolete">Paketbau Fortschritt:</translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1074"/>
+        <location filename="../oPB/core/processing.py" line="1128"/>
         <source>In progress:</source>
         <translation>In Bearbeitung:</translation>
     </message>
@@ -2738,74 +2754,89 @@ p, li { white-space: pre-wrap; }
         <translation>Paket öffnen (F2)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="508"/>
+        <location filename="../oPB/ui/startup.ui" line="411"/>
         <source>F2</source>
         <translation></translation>
     </message>
     <message>
         <location filename="../oPB/ui/startup.ui" line="331"/>
         <source>Recent files (F3)</source>
-        <translation>Zuletzt... (F3)</translation>
+        <translation type="obsolete">Zuletzt... (F3)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="536"/>
+        <location filename="../oPB/ui/startup.ui" line="535"/>
         <source>Extras</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="571"/>
+        <location filename="../oPB/ui/startup.ui" line="570"/>
         <source>Settings</source>
         <translation>Einstellungen</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="609"/>
+        <location filename="../oPB/ui/startup.ui" line="608"/>
         <source>Exit (Q)</source>
         <translation>Beenden (Q)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="650"/>
+        <location filename="../oPB/ui/startup.ui" line="649"/>
         <source>Depot Management (F9)</source>
         <translation>Depot Manager (F9)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="664"/>
+        <location filename="../oPB/ui/startup.ui" line="663"/>
         <source>F9</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="704"/>
+        <location filename="../oPB/ui/startup.ui" line="703"/>
         <source>Deploy client agent (F10)</source>
         <translation>Client Agent installieren (F10)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="759"/>
+        <location filename="../oPB/ui/startup.ui" line="717"/>
         <source>F10</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="786"/>
+        <location filename="../oPB/ui/startup.ui" line="782"/>
         <source>Job Scheduler (F4)</source>
         <translation>Zeitplaner (F4)</translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="800"/>
+        <location filename="../oPB/ui/startup.ui" line="796"/>
         <source>F4</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="623"/>
+        <location filename="../oPB/ui/startup.ui" line="622"/>
         <source>Q</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/ui/startup.ui" line="745"/>
+        <location filename="../oPB/ui/startup.ui" line="744"/>
         <source>Show Log</source>
         <translation>Zeige Log</translation>
     </message>
     <message>
         <location filename="../oPB/ui/startup.ui" line="494"/>
         <source>Import Package</source>
-        <translation>Paketdatei importieren</translation>
+        <translation type="obsolete">Paketdatei importieren</translation>
+    </message>
+    <message>
+        <location filename="../oPB/ui/startup.ui" line="331"/>
+        <source>Recent files</source>
+        <translation>Zuletzt...</translation>
+    </message>
+    <message>
+        <location filename="../oPB/ui/startup.ui" line="494"/>
+        <source>Import Package (F3)</source>
+        <translation>Paketdatei importieren (F3)</translation>
+    </message>
+    <message>
+        <location filename="../oPB/ui/startup.ui" line="507"/>
+        <source>F3</source>
+        <translation>F3</translation>
     </message>
 </context>
 </TS>